requirements.txt
# Core dependencies
streamlit>=1.22.0
chromadb>=0.4.18
google-generativeai>=0.3.0
nltk>=3.8.1
pandas>=1.5.3
numpy>=1.24.3
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.65.0